fix: move function
