+++ src/app/signupdetail/page.tsx
@@ -1,15 +1,14 @@
 "use client";
 
-import { detailCheck } from "@/action/auth";
 import { api } from "@/trpc/react";
 import { redirect } from "next/navigation";
 import { useEffect, useRef, useState } from "react";
 
-export default function SignUp() { // Participant
+export default function ParticipantSignUp() { 
   const { data, isSuccess } = api.auth.getAllUsers.useQuery();
   const [_,setError] = useState<string|null>(null)
   const { data: userData } = api.auth.me.useQuery();
-  const checkAKA = api.auth.isExistAKA.useMutation();
+  const checkAKA = api.auth.isAKAAlreadyExist.useMutation();
   useEffect(() => {
     if (userData) {
       redirect("/");
@@ -26,7 +25,7 @@ export default function SignUp() { // Participant
     }
 
     const formData = new FormData(formRef.current);
-    let err = await detailCheck(formData);
+    let err = await checkDetail(formData);
     if (!err) {
       const AKA = formData.get("username") as string
       const isAKAExist = await checkAKA.mutateAsync({ aka: AKA });
@@ -102,4 +101,31 @@ export default function SignUp() { // Participant
       </div>
     </main>
   );
+}
+
+async function checkDetail(formData:FormData){
+  try{
+    const aka = formData.get("username") as string | null;
+    const firstName = formData.get("firstName") as string | null;
+    const lastName = formData.get("lastName") as string | null;
+    const dateOfBirth = formData.get("dateOfBirth") as string | null;
+    const gender = formData.get("gender") as string | null;
+    
+    if(!aka){
+      throw new Error("Missing AKA")
+    }
+    if(!firstName || !lastName){
+      throw new Error("Missing firstName or lastName")
+    }
+    if(!gender){
+      throw new Error("Missing gender")
+    }
+    if(!dateOfBirth){
+      throw new Error("MissingBirthDate")
+    }
+    return null
+  }
+  catch (error){
+    return (error as Error).message
+  }
 }
